+++ desc/compute/_old.py
@@ -9,7 +9,6 @@ from orthax.legendre import leggauss
 
 from desc.backend import jit, jnp
 
-from ..batching import batch_map
 from ..integrals.bounce_integral import Bounce1D
 from ..integrals.quad_utils import (
     automorphism_sin,
@@ -29,43 +28,6 @@ _bounce1D_doc = {
     "surf_batch_size": _bounce_doc["surf_batch_size"],
     "quad": _bounce_doc["quad"],
 }
-
-
-def _compute(fun, fun_data, data, grid, num_pitch, surf_batch_size=1, simp=False):
-    """Compute Bounce1D integral quantity with ``fun``.
-
-    Parameters
-    ----------
-    fun : callable
-        Function to compute.
-    fun_data : dict[str, jnp.ndarray]
-        Data to provide to ``fun``. This dict will be modified.
-    data : dict[str, jnp.ndarray]
-        DESC data dict.
-    grid : Grid
-        Grid that can expand and compress.
-    num_pitch : int
-        Resolution for quadrature over velocity coordinate.
-    surf_batch_size : int
-        Number of flux surfaces with which to compute simultaneously.
-        Default is ``1``.
-    simp : bool
-        Whether to use an open Simpson rule instead of uniform weights.
-
-    """
-    for name in Bounce1D.required_names:
-        fun_data[name] = data[name]
-    for name in fun_data:
-        fun_data[name] = Bounce1D.reshape(grid, fun_data[name])
-    fun_data["pitch_inv"], fun_data["pitch_inv weight"] = Bounce1D.get_pitch_inv_quad(
-        grid.compress(data["min_tz |B|"]),
-        grid.compress(data["max_tz |B|"]),
-        num_pitch,
-        simp=simp,
-    )
-    out = batch_map(fun, fun_data, surf_batch_size)
-    assert out.ndim == 1
-    return grid.expand(out)
 
 
 @register_compute_fun(
@@ -140,14 +102,14 @@ def _epsilon_32_1D(params, transforms, profiles, data, **kwargs):
     grid = transforms["grid"].source_grid
     B0 = data["max_tz |B|"]
     data["old effective ripple 3/2"] = (
-        _compute(
+        Bounce1D.batch(
             eps_32,
             {"|grad(rho)|*kappa_g": data["|grad(rho)|"] * data["kappa_g"]},
             data,
             grid,
             num_pitch,
             surf_batch_size,
-            simp=True,
+            expand_out=True,
         )
         / data["fieldline length"]
         * (B0 * data["R0"] / data["<|grad(rho)|>"]) ** 2
@@ -288,7 +250,16 @@ def _Gamma_c_1D(params, transforms, profiles, data, **kwargs):
     }
     grid = transforms["grid"].source_grid
     data["old Gamma_c"] = (
-        _compute(Gamma_c, fun_data, data, grid, num_pitch, surf_batch_size)
+        Bounce1D.batch(
+            Gamma_c,
+            fun_data,
+            data,
+            grid,
+            num_pitch,
+            surf_batch_size,
+            simp=False,
+            expand_out=True,
+        )
         / data["fieldline length"]
     )
     return data
@@ -371,13 +342,15 @@ def _Gamma_c_Velasco_1D(params, transforms, profiles, data, **kwargs):
 
     grid = transforms["grid"].source_grid
     data["old Gamma_c Velasco"] = (
-        _compute(
+        Bounce1D.batch(
             Gamma_c,
             {"cvdrift0": data["cvdrift0"], "gbdrift": data["gbdrift"]},
             data,
             grid,
             num_pitch,
             surf_batch_size,
+            simp=False,
+            expand_out=True,
         )
         / data["fieldline length"]
     )