+++ CHANGELOG.md
@@ -1,13 +1,27 @@
 Changelog
 =========
 
+v0.16.1
+-------
+
+New Features
+
+- Method to plot Fourier spectrum of stream map in field line coordinates in ``Bounce2D``.
+- Utility method to compute integrals over phase space in batches is now added to the public API.
+
+Bug Fixes
+
+- [Improves convergence of stream map in ``Bounce2D``](https://github.com/PlasmaControl/DESC/pull/1919).
+
+v0.16.0
+-------
+
 New Features
 
 - New basis vector and metric elements derivatives in PEST coordinates and quantities useful for a global MHD stability solver.
 - Adds ``desc.external.TERPSICHORE`` objective for computing linear ideal MHD stability growth rates. This objective subclasses from ``ExternalObjective`` and requires access to the TERPSICHORE code, which is not included with DESC or its dependencies.
 - Adds ``docs/dev_guids/external_objectives.rst`` as a tutorial for how to use external objectives, with TERPSICHORE as an example using parallel processing.
 - Adds keyword argument `normalize` to plot_1d, plot_3d. `normalize` is a string to use for normalization.
-- New method to plot Fourier spectrum of stream map in field line coordinates in ``Bounce2D``.
 - Adds new linear objective `ShareParameters` which can enforce that the chosen parameters of two or more objects of the same type remain identical during an optimization. Potentially useful for flexible stellarator optimization, where one has two coilsets with the same geometry but differing currents, and attempts to optimize for two different stellarator equilibria.
 - Changes related to ``field_line_integrate``:
     - `field_line_integrate` now returns `diffrax.diffeqsolve.stats` and `diffrax.diffeqsolve.result` if the flag `return_aux` is set to True.
@@ -20,7 +34,6 @@ Bug Fixes
 
 - [Fixes straight field line equilibrium conversion](https://github.com/PlasmaControl/DESC/pull/1880).
 - ``desc.compat.rescale`` will now return ``ScaledProfile`` instances for most of its profiles, to fix a bug where improper scaling could occur for certain profile types.
-- [Improves convergence of stream map in ``Bounce2D``](https://github.com/PlasmaControl/DESC/pull/1919).
 - Now always use ``sym=False`` in the default grid for ``plot_fsa`` to ensure correct averages
 - Fixes bug that could lead extra compilation of jit-compiled functions that include `field_line_integrate`.
 - Fixes inaccurate normalizations scales that could be computed for certain equilibria which had m=1 n=0 R and m=-1 n=0 Z components much smaller than their actual average minor radius, see [GH issue](https://github.com/PlasmaControl/DESC/issues/1954)